refactor: created CurrentUserProvider to wrap all components with UserContext within layout.tsx, as well as changed redirection to refresh client side view
