+++ src/components/Login.tsx
@@ -5,11 +5,9 @@ import { authApi } from "@/utils/AuthApi";
 import { toast } from "react-toastify";
 import { setToken } from "@/utils/token";
 import { CurrentUserContext } from "@/contexts/CurrentUserContext";
-import { useRouter } from "next/navigation";
 
 export default function Login() {
   const userContext = useContext(CurrentUserContext);
-  const router = useRouter();
 
   const [data, setData] = useState({
     email: "" as string,
@@ -25,12 +23,14 @@ export default function Login() {
           userContext?.setCurrentUser(res.data);
           setToken(res.token);
           toast.success(res.message);
+
+          setTimeout(() => {
+            window.location.href = "/saved-news";
+          }, 1000);
         }
       })
-      .then(() => router.push("/saved-news"))
       .catch((err) => toast.error(err.message));
   };
-  console.log(userContext?.currentUser);
 
   const handleChange = (e: ChangeEvent) => {
     const { name, value } = e.target as HTMLInputElement;
